// ui/src/generated/rust-types/PersonOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Person summary (face recognition)
 */
export type PersonOutput = { id: string, name: string | null, faceCount: number, avatarPhotoId: string | null, avatarThumbnailPath: string | null, };